Fix memoised elements not updating on error object change
Presumably, the react-hook-form 'formState.errors' object's reference was not changing when its contents changed, so the memoised form inputs that relied on it as a dependency weren't displaying their errors when the value of their fields failed the validation in the Zod schema being used as a resolver. Now, the inputs just have the formState object itself as a dependency and this seems to fix the issue.