Adjust useEffect hook that updates variant state
The function that recalculates the variant state was being called whenever the product state was updated due to a useEffect hook using the product state as a dependency, but the variant state was already being set from within the function that also set the product state. This led to the variant state being updated twice every time the product state changed, so I am now storing the product state in a ref and am no longer monitoring the product state in that useEffect hook, which prevents the second update.