Change Root component's context's Header forceClose function
As it can be accessed from any descendant component, there may be some competition when trying to set whether or not the Header should be forced closed, so the function now requires an id as a second argument. This is stored within a Set in the Header component, and the Header will only be allowed to be open when the size of this Set is 0. This allows multiple sources to all reenable the Header's ability to open by removing this id from the Set, without having to worry if other components are also currently trying to keep it closed.