Fix issue with hook dependency
The hook was using the entire headerInfo state object as a dependency, which was causing the Header to be open when auto-scrolling to the top of the Review components' container, when the Header was already open. Now it just has the forceClose function as a dependency, which fixes the issue.