Restructure HTTP method API function types
The abortController is now within the same props object as 'params' and 'body' fields, which should mean it shouldn't need to be defined in order to use additional args in the function in cases where it's not applicable. I also created a FuncResponseObject type to define a consistent structure for the response object (status, message, data) across all HTTP methods' API function types.