Overhaul functionality of Quantity component
I wanted to allow the user to type any integer value, both positive and negative, into the input, then clamp the value within the bounds of 'min' and 'max' (or the default value or base value depending on which props are defined) on blur. To do this, using a type='number' input was preventing the real value from being accessible in onInput, as the value was being sanitised before it was accessible in the form event object. To work around this, I have to use type='text', but I will attempt to improve the semantics of the input to assert it is numeric, for accessibility purposes.