Force state change for internal variables when props are mutated
While a subscription's data is being awaited, the ScheduleModal component's 'data' prop is provided by some default subscriptions that exist to provide layout to skeleton loaders. When the subscription's data is received, the internal state values for these fields that were being generated when the component mounted weren't being updated, so the 'defaultValue' prop passed to the child Quantity component was coming from the default data, not the real data. Forcing a state change when the subscription's data changes (i.e. - when the real data is received) fixes this issue.