Ensure undefined field values are correctly normalised
I had an issue whereby deleting a field's value, causing it to become 'undefined' due to my onChange handler logic, was creating a disparity between the actual value of that field (undefined) and the value actually being shown in the input to the user. Normalising undefined values causes the input to correctly clear when the input's value is deleted.