Set 'transitionDuration' prop on Collapse component to 0
The Collapse component's transition seems to calculate height and then animate based on that height value, but this doesn't play well with child elements with inconsistent height. For example, the list of cart items within the component is scrollable when the height of the component and button would exceed the available space in the viewport; this causes a jump at the end of the transition when the height is properly calculated and the component realises it needs to shrink to fit the available space.