Fix layout shift issue on page load by using media queries
Mantine's useMatches hook was causing an issue where briefly on page load it wasn't recognising the real width of the viewport and returning an incorrect value. This was causing the wrong 'layout' prop value to be passed to the child CartSummary component, resulting in the wrong page layout on page load. Switching over to media queries and 'hiddenFrom'/'visibleFrom' Mantine component props for the CheckoutContent component fixes this issue.