Mock Routes module in AccountDetails component's test suite
Some warnings were being thrown about mocking the components, importing them into the Routes module, then using that in the component. I think it was something to do with a disparity between the evaluation times of the mocks and importing the Routes module into the AccountDetails component's module, so just mocking the entire Routes module fixes this issue.