Update fetcher function to use status and message from response
Previously, I was adding 'status' and 'message' fields to the response object's body along with the 'data' field, but my backend now only has the data itself stored in the body. The fetcher function now also correctly catches errors thrown when trying to parse the response object's body's JSON when there is nothing to parse.