Make 'data' and 'errors' objects contain valid data
Before, the useQueryContexts hook returned 'data' and 'errors' object fields that had all provided contexts' 'name' field values as keys, but the values for those fields didn't have to be defined. Now, only successful responses with valid data will add that context's 'name' field and data to the 'data' object, and only unsuccessful responses will add that context's 'name field and the error object from the response (or the response message in lieu of an error object) to the 'errors' object.