Add constraint to product_variant_attributes table
This constraint enforces a specified attribute_id must be present in the product_attribute_orders table for the specified product. I had to introduce some redundancy in the table by adding the product_id field (it's already implied by the product_variant_id field which references a product_variant which must only reference one product), but it's needed for the constraint, and adding another constraint enforcing the link between the product_id and product_variant_id prevents adding attributes to a product/variant combination that doesn't exist.