Write JSX and styles for listing select-type category filters
I've called it 'select' in the database schema, but in reality I'm actually using radio buttons for now. It serves the same purpose by only allowing the user to select one of the values at any given time; the visual formatting is just different. I may implement a select or collapsible container in the future for the radio buttons for filters with large numbers of options.